+++ packages/cutwater-build-core/src/index.ts
@@ -302,7 +302,7 @@ const generateGulpError = (err: Error): Error => {
     };
     logger.markErrorAsWritten(rval);
   }
-  return rval;
+  return rval as Error;
 };
 
 function executeTask(taskExecutable: ExecutableTask<unknown>, localContext: BuildContext): Promise<void> {
@@ -397,5 +397,8 @@ task('prettier', prettier);
 
 export const jestIntegration: JestTask = new JestTask();
 jestIntegration.name = 'jest-integration';
-jestIntegration.setConfig({ isEnabled: true, testMatch: ['<rootDir>/lib/**/*.(integ).js?(x)'] });
+jestIntegration.setConfig({
+  isEnabled: true,
+  options: { ...jestIntegration.config.options, testMatch: ['<rootDir>/src/**/*.(integ).(ts|js)?(x)'] },
+});
 task('jest-integration', jestIntegration);