+++ packages/cutwater-build-webpack/src/tasks/webpack.config.ts
@@ -4,52 +4,56 @@ import * as TerserPlugin from 'terser-webpack-plugin';
 import * as Webpack from 'webpack';
 
 const buildConfig: BuildConfig = getConfig();
-const isProduction: boolean = buildConfig.production;
-
 const packageJSON: { name: string } = IOUtils.readJSONSyncSafe('./package.json');
 
-const webpackConfiguration: Webpack.Configuration = {
-  mode: isProduction ? 'production' : 'development',
-  context: __dirname,
-  devtool: isProduction ? undefined : 'source-map',
+const webpackConfiguration = (env, options): Webpack.Configuration => {
+  const isProduction: boolean = options.mode === 'production';
 
-  entry: {
-    [packageJSON.name]: path.join(__dirname, buildConfig.libFolder, 'index.js'),
-  },
+  const minimizer: any[] = [];
+  if (isProduction) {
+    minimizer.push(
+      new TerserPlugin({
+        parallel: true,
+        sourceMap: true,
+        include: /\.min\.js$/,
+        terserOptions: {
+          ecma: 6,
+        },
+      }),
+    );
+  }
 
-  output: {
-    libraryTarget: 'umd',
-    path: path.join(__dirname, buildConfig.distFolder),
-    filename: `[name]${isProduction ? '.min' : ''}.js`,
-    globalObject: 'this',
-  },
+  return {
+    mode: isProduction ? 'production' : 'development',
+    context: __dirname,
+    devtool: isProduction ? undefined : 'source-map',
 
-  externals: {
-    react: {
-      amd: 'react',
-      commonjs: 'react',
+    entry: {
+      [packageJSON.name]: path.join(__dirname, buildConfig.libFolder, 'index.js'),
     },
-    'react-dom': {
-      amd: 'react-dom',
-      commonjs: 'react-dom',
-    },
-  } as any,
 
-  optimization: {
-    minimizer: [],
-  },
-};
-if (isProduction && webpackConfiguration.optimization && webpackConfiguration.optimization.minimizer) {
-  webpackConfiguration.optimization.minimizer.push(
-    new TerserPlugin({
-      parallel: true,
-      sourceMap: true,
-      include: /\.min\.js$/,
-      terserOptions: {
-        ecma: 6,
+    output: {
+      libraryTarget: 'umd',
+      path: path.join(__dirname, buildConfig.distFolder),
+      filename: `[name]${isProduction ? '.min' : ''}.js`,
+      globalObject: 'this',
+    },
+
+    externals: {
+      react: {
+        amd: 'react',
+        commonjs: 'react',
       },
-    }),
-  );
-}
+      'react-dom': {
+        amd: 'react-dom',
+        commonjs: 'react-dom',
+      },
+    } as any,
+
+    optimization: {
+      minimizer,
+    },
+  };
+};
 
 module.exports = webpackConfiguration;