+++ CHANGELOG.md
@@ -3,6 +3,127 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.24.1 (2021-03-03)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+# 0.24.0 (2021-02-15)
+
+
+### Features
+
+* improve CI/CD workflow ([6a6ada5](https://github.com/CodificationOrg/cutwater/commit/6a6ada51b261da2296fc9a4628be3abbdc3021f8))
+
+
+
+
+
+## [0.23.4](https://github.com/CodificationOrg/cutwater/compare/v0.23.3...v0.23.4) (2021-01-07)
+
+
+### Bug Fixes
+
+* adjust method of setting mode for webpack ([e454662](https://github.com/CodificationOrg/cutwater/commit/e4546628d69243a4ef8de63ba6f2f9dfa1f44e2a))
+
+
+
+
+
+
+## [0.23.3](https://github.com/CodificationOrg/cutwater/compare/v0.23.2...v0.23.3) (2021-01-07)
+
+
+### Bug Fixes
+
+* ensure prod flag is passed to webpack ([44b8601](https://github.com/CodificationOrg/cutwater/commit/44b860114fb5ebfc3df002c80a53e80c2a61580d))
+
+
+
+
+
+
+## [0.23.2](https://github.com/CodificationOrg/cutwater/compare/v0.23.1...v0.23.2) (2020-12-28)
+
+
+### Bug Fixes
+
+* correct failure to clean some tsc outputs ([1d36ef6](https://github.com/CodificationOrg/cutwater/commit/1d36ef645f3163c263c262e9d835aa51971c9b35))
+
+
+
+
+
+
+## [0.23.1](https://github.com/CodificationOrg/cutwater/compare/v0.23.0...v0.23.1) (2020-12-26)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+
+# [0.23.0](https://github.com/CodificationOrg/cutwater/compare/v0.22.1...v0.23.0) (2020-12-25)
+
+
+### Features
+
+* move webpack to native calls ([b643731](https://github.com/CodificationOrg/cutwater/commit/b64373134d7ddfa15ba16581bb53630e02df9550))
+
+
+
+
+
+
+## [0.22.1](https://github.com/CodificationOrg/cutwater/compare/v0.22.0...v0.22.1) (2020-12-25)
+
+
+### Bug Fixes
+
+* update param refs to use options ([1adb498](https://github.com/CodificationOrg/cutwater/commit/1adb49887e7cfff83d427ac9c112fd4d69401186))
+
+
+
+
+
+
+# [0.22.0](https://github.com/CodificationOrg/cutwater/compare/v0.21.4...v0.22.0) (2020-12-25)
+
+
+### Features
+
+* convert to using native jest call ([6374658](https://github.com/CodificationOrg/cutwater/commit/6374658ad9fbba60f461cffdad4cbc891a1470f5))
+* convert TscTask to use native tsc call ([a2d77ad](https://github.com/CodificationOrg/cutwater/commit/a2d77ad4a48b2d3fc0a62fd7ecb47867a7d59571))
+
+
+
+
+
+
+## [0.21.4](https://github.com/CodificationOrg/cutwater/compare/v0.21.3...v0.21.4) (2020-12-23)
+
+
+### Bug Fixes
+
+* correct problems caused by new versions ([c6ae7a5](https://github.com/CodificationOrg/cutwater/commit/c6ae7a57fda5d5ad95c5ee4ad8596cdca9b7f91a))
+
+
+
+
+
+
+## [0.21.3](https://github.com/CodificationOrg/cutwater/compare/v0.21.2...v0.21.3) (2020-10-31)
+
+**Note:** Version bump only for package root
+
+
+
+
+
 ## [0.21.2](https://github.com/CodificationOrg/cutwater/compare/v0.21.1...v0.21.2) (2020-08-16)
 
 **Note:** Version bump only for package root